fix blinking animation
