+++ app/src/main/java/com/canli/oya/traininventoryroom/utils/Extensions.kt
@@ -2,10 +2,19 @@ package com.canli.oya.traininventoryroom.utils
 
 import android.app.Activity
 import android.content.Context
+import android.graphics.drawable.Animatable2
+import android.graphics.drawable.AnimatedVectorDrawable
+import android.graphics.drawable.Drawable
+import android.os.Build
 import android.view.Gravity
+import android.view.MenuItem
 import android.view.inputmethod.InputMethodManager
 import android.widget.Toast
+import androidx.annotation.DrawableRes
 import androidx.annotation.StringRes
+import com.canli.oya.traininventoryroom.R
+import com.canli.oya.traininventoryroom.ui.base.setMenuIcon
+import timber.log.Timber
 
 private var toast : Toast? = null
 
@@ -48,5 +57,24 @@ fun Activity.clearFocusAndHideKeyboard() {
     focusedView?.run {
         clearFocus()
         imm.hideSoftInputFromWindow(this.windowToken, 0)
+    }
+}
+
+fun MenuItem.blinkAddMenuItem(@DrawableRes iconToSet: Int) {
+    Timber.d("blinkAddMenuItem is called")
+    if (Build.VERSION.SDK_INT >= 23) {
+        setIcon(R.drawable.avd_blinking_plus)
+        val blinkingAnim = this.icon as? AnimatedVectorDrawable
+        blinkingAnim?.clearAnimationCallbacks()
+        blinkingAnim?.registerAnimationCallback(object : Animatable2.AnimationCallback() {
+            override fun onAnimationStart(drawable: Drawable) {
+                Timber.d("blink animation started")
+            }
+
+            override fun onAnimationEnd(drawable: Drawable) {
+                setMenuIcon(iconToSet)
+            }
+        })
+        blinkingAnim?.start()
     }
 }
